+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2023-10-02
+
+### Added
+
+- Toggling eventID in the output
+- Filter out events that have no username or issuer ARN
+
 ## [0.0.2] - 2023-08-16
 
 ### Added